@@ -19,17 +19,16 @@ publish_to: "none" # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=3.0.0 <4.0.0"
 
 dependencies:
-  appwrite: ^5.0.0
+  appwrite: ^11.0.1
   cupertino_icons: ^1.0.2
   flutter:
     sdk: flutter
-  flutter_lints: ^2.0.1
-  flutter_riverpod: ^1.0.1
+  flutter_lints: ^3.0.1
+  flutter_riverpod: ^2.4.9
   font_awesome_flutter: ^10.1.0
-  google_sign_in: ^5.0.7
 
 dev_dependencies:
   flutter_test:
