@@ -6,8 +6,8 @@ class ErrorScreen extends StatelessWidget {
   const ErrorScreen(
     this.e,
     this.trace, {
-    Key? key,
-  }) : super(key: key);
+    super.key,
+  });
 
   @override
   Widget build(BuildContext context) {
